requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Django==4.2.10
django-bootstrap-v5==1.0.11
django-extensions==3.2.3
django-fontawesome-5==1.0.18
django-widget-tweaks==1.5.0
gunicorn==21.2.0
packaging==24.0
python-decouple==3.8
soupsieve==2.5
sqlparse==0.4.4
django-role-permissions-3.2.0
